[test] move interpreters tests to tests/it
