+++ query/tests/it/interpreters/interpreter_grant_privilege.rs
@@ -16,46 +16,43 @@ use common_base::tokio;
 use common_exception::Result;
 use common_management::UserInfo;
 use common_meta_types::AuthType;
+use common_meta_types::UserPrivilege;
 use common_planners::*;
+use databend_query::interpreters::*;
+use databend_query::sql::PlanParser;
 use futures::stream::StreamExt;
 use pretty_assertions::assert_eq;
 
-use crate::interpreters::*;
-use crate::sql::*;
-
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_alter_user_interpreter() -> Result<()> {
+async fn test_grant_privilege_interpreter() -> Result<()> {
     common_tracing::init_default_ut_tracing();
 
     let ctx = crate::tests::try_create_context()?;
     let name = "test";
     let hostname = "localhost";
     let password = "test";
-    let new_password = "password";
     let user_info = UserInfo::new(
         name.to_string(),
         hostname.to_string(),
         Vec::from(password),
         AuthType::PlainText,
     );
+    assert_eq!(user_info.privileges, UserPrivilege::empty());
     let user_mgr = ctx.get_sessions_manager().get_user_manager();
     user_mgr.add_user(user_info).await?;
 
-    let old_user = user_mgr.get_user(name, hostname).await?;
-    assert_eq!(old_user.password, Vec::from(password));
-
-    let test_query = format!(
-        "ALTER USER '{}'@'{}' IDENTIFIED BY '{}'",
-        name, hostname, new_password
-    );
-
-    if let PlanNode::AlterUser(plan) = PlanParser::parse(&test_query, ctx.clone()).await? {
-        let executor = AlterUserInterpreter::try_create(ctx, plan.clone())?;
-        assert_eq!(executor.name(), "AlterUserInterpreter");
+    let test_query = format!("GRANT ALL ON * TO '{}'@'{}'", name, hostname);
+    if let PlanNode::GrantPrivilege(plan) = PlanParser::parse(&test_query, ctx.clone()).await? {
+        let executor = GrantPrivilegeInterpreter::try_create(ctx, plan.clone())?;
+        assert_eq!(executor.name(), "GrantPrivilegeInterpreter");
         let mut stream = executor.execute(None).await?;
         while let Some(_block) = stream.next().await {}
         let new_user = user_mgr.get_user(name, hostname).await?;
-        assert_eq!(new_user.password, Vec::from(new_password))
+        assert_eq!(new_user.privileges, {
+            let mut privileges = UserPrivilege::empty();
+            privileges.set_all_privileges();
+            privileges
+        })
     } else {
         panic!()
     }
