+++ query/tests/it/interpreters/interpreter_select.rs
@@ -15,11 +15,10 @@
 use common_base::tokio;
 use common_exception::Result;
 use common_planners::*;
+use databend_query::interpreters::*;
+use databend_query::sql::*;
 use futures::TryStreamExt;
 use pretty_assertions::assert_eq;
-
-use crate::interpreters::*;
-use crate::sql::*;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
 async fn test_select_interpreter() -> Result<()> {
