+++ query/tests/it/interpreters/interpreter_table_create.rs
@@ -15,41 +15,27 @@
 use common_base::tokio;
 use common_exception::Result;
 use common_planners::*;
+use databend_query::interpreters::*;
 use futures::stream::StreamExt;
-use pretty_assertions::assert_eq;
 
-use crate::interpreters::*;
 use crate::tests::parse_query;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_setting_interpreter() -> Result<()> {
+async fn test_create_table_interpreter() -> Result<()> {
     let ctx = crate::tests::try_create_context()?;
 
-    if let PlanNode::SetVariable(plan) = parse_query("SET max_block_size=1", &ctx)? {
-        let executor = SettingInterpreter::try_create(ctx, plan)?;
-        assert_eq!(executor.name(), "SettingInterpreter");
+    static TEST_CREATE_QUERY: &str = "\
+        CREATE TABLE default.a(\
+            a bigint, b int, c varchar(255), d smallint, e Date\
+        ) Engine = Null\
+    ";
 
-        let mut stream = executor.execute(None).await?;
+    if let PlanNode::CreateTable(plan) = parse_query(TEST_CREATE_QUERY, &ctx)? {
+        let interpreter = CreateTableInterpreter::try_create(ctx, plan.clone())?;
+        let mut stream = interpreter.execute(None).await?;
         while let Some(_block) = stream.next().await {}
     } else {
         panic!()
-    }
-
-    Ok(())
-}
-
-#[tokio::test(flavor = "multi_thread", worker_threads = 1)]
-async fn test_setting_interpreter_error() -> Result<()> {
-    let ctx = crate::tests::try_create_context()?;
-
-    if let PlanNode::SetVariable(plan) = parse_query("SET xx = 1", &ctx)? {
-        let executor = SettingInterpreter::try_create(ctx, plan)?;
-        if let Err(e) = executor.execute(None).await {
-            let expect = "Code: 20, displayText = Unknown variable: \"xx\".";
-            assert_eq!(expect, format!("{}", e));
-        } else {
-            panic!();
-        }
     }
 
     Ok(())
